renamed e2e-test to e2e-smoke in circleci as well as renamed cypress/e2e to cypress/e2e-smoke
