+++ front-end/cypress/extended/contacts/contacts.list.cy.ts
@@ -1,10 +1,10 @@
-import { Initialize } from '../../e2e/pages/loginPage';
-import { ContactListPage } from '../../e2e/pages/contactListPage';
+import { Initialize } from '../../e2e-smoke/pages/loginPage';
+import { ContactListPage } from '../../e2e-smoke/pages/contactListPage';
 import { ContactsHelpers } from './contacts.helpers';
-import { PageUtils } from '../../e2e/pages/pageUtils';
-import { defaultFormData as contactFormData } from '../../e2e/models/ContactFormModel';
-import { makeContact } from '../../e2e/requests/methods';
-import { Individual_A_A, MockContact } from '../../e2e/requests/library/contacts';
+import { PageUtils } from '../../e2e-smoke/pages/pageUtils';
+import { defaultFormData as contactFormData } from '../../e2e-smoke/models/ContactFormModel';
+import { makeContact } from '../../e2e-smoke/requests/methods';
+import { Individual_A_A, MockContact } from '../../e2e-smoke/requests/library/contacts';
 
 describe('Contacts List (/contacts)', () => {
   beforeEach(() => {
